Match clean up

+++ examples/src/main/java/me/breidenbach/util/examples/MatchExample.java
@@ -52,7 +52,7 @@ class MatchExample {
 
         final String answer = match(pi,
                 typeCase(Integer.class, j -> "Not today"),
-                matchCase(j -> (Double)j == 3.1415927, j-> "Eureka!"));
+                matchCase(j -> j == 3.1415927, j-> "Eureka!"));
 
         if (!answer.equals("Eureka!")) throw new RuntimeException("Sunk");
